+++ src/converters.ts
@@ -1,5 +1,5 @@
 import { IObservableArray, observable } from "mobx";
-import { IModelType } from "mobx-state-tree";
+import { IModelType, IAnyModelType, Instance } from "mobx-state-tree";
 import {
   CONVERSION_ERROR,
   ConversionResponse,
@@ -152,26 +152,45 @@ const stringArray = new Converter<string[], IObservableArray<string>>({
   }
 });
 
-// this works with string converters and also with models
 function maybe<R, V>(
   converter: StringConverter<V>
-): IConverter<string, V | null>;
-function maybe<R>(converter: IConverter<R, R>): IConverter<R | null, R | null>;
+): IConverter<string, V | undefined>;
+function maybe<M>(
+  converter: IConverter<M | null, M | undefined>
+): IConverter<M | null, M | undefined>;
 function maybe<R, V>(
-  converter: Converter<string, V> | IConverter<R, R>
-): IConverter<string, V | null> | IConverter<R | null, R | null> {
+  converter: IConverter<R, V>
+): IConverter<string, V | undefined> | IConverter<R | null, V | undefined> {
   if (converter instanceof StringConverter || converter instanceof Decimal) {
-    return new StringMaybe(converter);
+    return new StringMaybe(converter, undefined);
   }
-  return maybeModel(converter as IConverter<R, R>);
+  return maybeModel(converter, null, undefined) as IConverter<
+    R | null,
+    V | undefined
+  >;
 }
 
-class StringMaybe<V> implements IConverter<string, V | null> {
+function maybeNull<R, V>(
+  converter: StringConverter<V>
+): IConverter<string, V | null>;
+function maybeNull<M>(
+  converter: IConverter<M | null, M | null>
+): IConverter<M | null, M | null>;
+function maybeNull<R, V>(
+  converter: IConverter<R, V>
+): IConverter<string, V | null> | IConverter<R | null, V | null> {
+  if (converter instanceof StringConverter || converter instanceof Decimal) {
+    return new StringMaybe(converter, null);
+  }
+  return maybeModel(converter, null, null) as IConverter<R | null, V | null>;
+}
+
+class StringMaybe<V, RE, VE> implements IConverter<string, V | VE> {
   emptyRaw: string;
   defaultControlled = controlled.value;
   neverRequired = false;
 
-  constructor(public converter: IConverter<string, V>) {
+  constructor(public converter: IConverter<string, V>, public emptyValue: VE) {
     this.emptyRaw = "";
   }
 
@@ -179,57 +198,63 @@ class StringMaybe<V> implements IConverter<string, V | null> {
     return raw.trim();
   }
 
-  async convert(raw: string): Promise<ConversionResponse<V | null>> {
+  async convert(raw: string): Promise<ConversionResponse<V | VE>> {
     if (raw.trim() === "") {
-      return new ConversionValue(null);
+      return new ConversionValue(this.emptyValue);
     }
     return this.converter.convert(raw);
   }
 
-  render(value: V | null): string {
-    if (value === null) {
+  render(value: V | VE): string {
+    if (value === this.emptyValue) {
       return "";
     }
-    return this.converter.render(value);
+    return this.converter.render(value as V);
   }
 }
 
-class Model<M> implements IConverter<M | null, M> {
-  emptyRaw: M | null;
+class Model<M, RE> implements IConverter<Instance<M> | RE, Instance<M>> {
+  emptyRaw: Instance<M> | RE;
   defaultControlled: Controlled;
   neverRequired = false;
 
-  constructor(model: IModelType<any, M>) {
-    this.emptyRaw = null;
+  constructor(model: M, emptyRaw: RE) {
+    this.emptyRaw = emptyRaw;
     this.defaultControlled = controlled.object;
   }
-  preprocessRaw(raw: M): M {
+  preprocessRaw(raw: Instance<M>): Instance<M> {
     return raw;
   }
 
-  async convert(raw: M | null): Promise<ConversionResponse<M>> {
-    if (raw === null) {
+  async convert(
+    raw: Instance<M> | RE
+  ): Promise<ConversionResponse<Instance<M>>> {
+    if (raw === this.emptyRaw) {
       return CONVERSION_ERROR;
     }
-    return new ConversionValue(raw);
+    return new ConversionValue(raw as Instance<M>);
   }
 
-  render(value: M): M {
+  render(value: Instance<M>): Instance<M> {
     return value;
   }
 }
 
-function model<M>(model: IModelType<any, M>): IConverter<M | null, M> {
-  return new Model(model);
+function model<M extends IAnyModelType>(
+  model: M
+): IConverter<Instance<M> | null, Instance<M>> {
+  return new Model(model, null);
 }
 
-function maybeModel<M>(
-  converter: IConverter<M, M>
-): IConverter<M | null, M | null> {
+function maybeModel<M, RE, VE>(
+  converter: IConverter<M | RE, M | VE>,
+  emptyRaw: RE,
+  emptyValue: VE
+): IConverter<M | RE, M | VE> {
   return new Converter({
-    emptyRaw: null,
-    convert: identity,
-    render: identity,
+    emptyRaw: emptyRaw,
+    convert: (r: M | RE) => (r !== emptyRaw ? (r as M) : emptyValue),
+    render: (v: M | VE) => (v !== emptyValue ? (v as M) : emptyRaw),
     defaultControlled: controlled.object
   });
 }
@@ -248,6 +273,7 @@ export const converters = {
   boolean,
   stringArray,
   maybe,
+  maybeNull,
   model,
   object
 };