+++ test/derived.test.ts
@@ -1,9 +1,9 @@
 import { configure, IReactionDisposer } from "mobx";
-import { types } from "mobx-state-tree";
+import { types, Instance } from "mobx-state-tree";
 import { Field, Form, RepeatingForm, converters } from "../src";
 
-// "strict" leads to trouble during initialization.
-configure({ enforceActions: true });
+// "always" leads to trouble during initialization.
+configure({ enforceActions: "observed" });
 
 // a way to wait for all reactions to have been resolved
 function resolveReactions() {
@@ -29,7 +29,7 @@ test("calculated", async () => {
 
   const form = new Form(M, {
     calculated: new Field(converters.number, {
-      derived: (node: typeof M.Type) => node.sum()
+      derived: (node: Instance<typeof M>) => node.sum()
     }),
     a: new Field(converters.number),
     b: new Field(converters.number)